Improving simple stroke comparisons. Deleting logs, old code.

+++ WriteJapanese/app/src/test/java/dmeeuwis/nakama/kanjidraw/SimpleDrawingComparatorTest.java
@@ -10,25 +10,25 @@ public class SimpleDrawingComparatorTest {
 
     @Test
     public void testMisorderedStrokesOK(){
-        findMatch(new int[][]{
-                new int[]{1, 0, 2},
-                new int[]{0, 1, 2},
-                new int[]{1, 1, 0}});
+        findMatch(new double[][]{
+                new double[]{1, 0, 2},
+                new double[]{0, 1, 2},
+                new double[]{1, 1, 0}});
 
-        findMatch(new int[][]{
-                new int[]{0, 1, 2},
-                new int[]{1, 2, 0},
-                new int[]{1, 0, 1}});
+        findMatch(new double[][]{
+                new double[]{0, 1, 2},
+                new double[]{1, 2, 0},
+                new double[]{1, 0, 1}});
 
         // taken from drawing hiragana mo with bad stroke order
-        findMatch(new int[][]{
-                new int[]{0, 1, 0},
-                new int[]{0, 0, 0},
-                new int[]{0, 0, 1}});
+        findMatch(new double[][]{
+                new double[]{0, 1, 0},
+                new double[]{0, 0, 0},
+                new double[]{0, 0, 1}});
     }
 
 
-    private static void findMatch(int[][] m){
+    private static void findMatch(double[][] m){
         List<StrokeResult> res = SimpleDrawingComparator.findBestPairings(m);
         for (StrokeResult r : res) {
             assertEquals("All strokes found a match", 0, r.score);
